--- src/test/java/com/pivovarit/collectors/ExceptionPropagationTest.java
@@ -1,265 +1,0 @@
-package com.pivovarit.collectors;
-
-import org.junit.jupiter.api.Test;
-
-import java.util.ArrayList;
-import java.util.concurrent.CompletionException;
-import java.util.stream.IntStream;
-
-import static com.pivovarit.collectors.ParallelCollectors.inParallelToCollection;
-import static com.pivovarit.collectors.ParallelCollectors.inParallelToList;
-import static com.pivovarit.collectors.ParallelCollectors.inParallelToSet;
-import static com.pivovarit.collectors.ParallelCollectors.supplier;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-/**
- * @author Grzegorz Piwowarek
- */
-class ExceptionPropagationTest extends ExecutorAwareTest {
-
-    @Test
-    void shouldCollectToCollectionAndNotSwallowException() {
-        // given
-        executor = threadPoolExecutor(10);
-
-        assertThatThrownBy(() -> {
-            IntStream.range(0, 10).boxed()
-              .map(i -> supplier(() -> {
-                  if (i == 7) {
-                      throw new IllegalArgumentException();
-                  } else {
-                      return i;
-                  }
-              }))
-              .collect(inParallelToCollection(ArrayList::new, executor, 10))
-              .join();
-        })
-          .isInstanceOf(CompletionException.class)
-          .hasCauseExactlyInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    void shouldCollectToCollectionAndNotSwallowExceptionUnbounded() {
-        // given
-        executor = threadPoolExecutor(10);
-
-        assertThatThrownBy(() -> {
-            IntStream.range(0, 10).boxed()
-              .map(i -> supplier(() -> {
-                  if (i == 7) {
-                      throw new IllegalArgumentException();
-                  } else {
-                      return i;
-                  }
-              }))
-              .collect(inParallelToCollection(ArrayList::new, executor))
-              .join();
-        })
-          .isInstanceOf(CompletionException.class)
-          .hasCauseExactlyInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    void shouldCollectToListAndNotSwallowException() {
-        // given
-        executor = threadPoolExecutor(10);
-
-        assertThatThrownBy(() -> {
-            IntStream.range(0, 10).boxed()
-              .map(i -> supplier(() -> {
-                  if (i == 7) {
-                      throw new IllegalArgumentException();
-                  } else {
-                      return i;
-                  }
-              }))
-              .collect(inParallelToList(executor, 10))
-              .join();
-        })
-          .isInstanceOf(CompletionException.class)
-          .hasCauseExactlyInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    void shouldCollectToListAndNotSwallowExceptionUnbounded() {
-        // given
-        executor = threadPoolExecutor(10);
-
-        assertThatThrownBy(() -> {
-            IntStream.range(0, 10).boxed()
-              .map(i -> supplier(() -> {
-                  if (i == 7) {
-                      throw new IllegalArgumentException();
-                  } else {
-                      return i;
-                  }
-              }))
-              .collect(inParallelToList(executor))
-              .join();
-        })
-          .isInstanceOf(CompletionException.class)
-          .hasCauseExactlyInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    void shouldCollectToSetAndNotSwallowException() {
-        // given
-        executor = threadPoolExecutor(10);
-
-        assertThatThrownBy(() -> {
-            IntStream.range(0, 10).boxed()
-              .map(i -> supplier(() -> {
-                  if (i == 7) {
-                      throw new IllegalArgumentException();
-                  } else {
-                      return i;
-                  }
-              }))
-              .collect(inParallelToSet(executor, 10))
-              .join();
-        })
-          .isInstanceOf(CompletionException.class)
-          .hasCauseExactlyInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    void shouldCollectToSetAndNotSwallowExceptionUnbounded() {
-        // given
-        executor = threadPoolExecutor(10);
-
-        assertThatThrownBy(() -> {
-            IntStream.range(0, 10).boxed()
-              .map(i -> supplier(() -> {
-                  if (i == 7) {
-                      throw new IllegalArgumentException();
-                  } else {
-                      return i;
-                  }
-              }))
-              .collect(inParallelToSet(executor))
-              .join();
-        })
-          .isInstanceOf(CompletionException.class)
-          .hasCauseExactlyInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    void shouldCollectToCollectionMappingAndNotSwallowException() {
-        // given
-        executor = threadPoolExecutor(10);
-
-        assertThatThrownBy(() -> {
-            IntStream.range(0, 10).boxed()
-              .collect(inParallelToCollection(i -> {
-                  if (i == 7) {
-                      throw new IllegalArgumentException();
-                  } else {
-                      return i;
-                  }
-              }, ArrayList::new, executor, 10))
-              .join();
-        })
-          .isInstanceOf(CompletionException.class)
-          .hasCauseExactlyInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    void shouldCollectToCollectionMappingAndNotSwallowExceptionUnbounded() {
-        // given
-        executor = threadPoolExecutor(10);
-
-        assertThatThrownBy(() -> {
-            IntStream.range(0, 10).boxed()
-              .collect(inParallelToCollection(i -> {
-                  if (i == 7) {
-                      throw new IllegalArgumentException();
-                  } else {
-                      return i;
-                  }
-              }, ArrayList::new, executor))
-              .join();
-        })
-          .isInstanceOf(CompletionException.class)
-          .hasCauseExactlyInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    void shouldCollectToListMappingAndNotSwallowException() {
-        // given
-        executor = threadPoolExecutor(10);
-
-        assertThatThrownBy(() -> {
-            IntStream.range(0, 10).boxed()
-              .collect(inParallelToList(i -> {
-                  if (i == 7) {
-                      throw new IllegalArgumentException();
-                  } else {
-                      return i;
-                  }
-              }, executor, 10))
-              .join();
-        })
-          .isInstanceOf(CompletionException.class)
-          .hasCauseExactlyInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    void shouldCollectToListMappingAndNotSwallowExceptionUnbounded() {
-        // given
-        executor = threadPoolExecutor(10);
-
-        assertThatThrownBy(() -> {
-            IntStream.range(0, 10).boxed()
-              .collect(inParallelToList(i -> {
-                  if (i == 7) {
-                      throw new IllegalArgumentException();
-                  } else {
-                      return i;
-                  }
-              }, executor))
-              .join();
-        })
-          .isInstanceOf(CompletionException.class)
-          .hasCauseExactlyInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    void shouldCollectToSetMappingAndNotSwallowException() {
-        // given
-        executor = threadPoolExecutor(10);
-
-        assertThatThrownBy(() -> {
-            IntStream.range(0, 10).boxed()
-              .collect(inParallelToSet(i -> {
-                  if (i == 7) {
-                      throw new IllegalArgumentException();
-                  } else {
-                      return i;
-                  }
-              }, executor, 10))
-              .join();
-        })
-          .isInstanceOf(CompletionException.class)
-          .hasCauseExactlyInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    void shouldCollectToSetMappingAndNotSwallowExceptionUnbounded() {
-        // given
-        executor = threadPoolExecutor(10);
-
-        assertThatThrownBy(() -> {
-            IntStream.range(0, 10).boxed()
-              .collect(inParallelToSet(i -> {
-                  if (i == 7) {
-                      throw new IllegalArgumentException();
-                  } else {
-                      return i;
-                  }
-              }, executor))
-              .join();
-        })
-          .isInstanceOf(CompletionException.class)
-          .hasCauseExactlyInstanceOf(IllegalArgumentException.class);
-    }
-}
